requirements.txt
autopep8==1.6.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
idna==3.3
Pillow==9.1.1
pycodestyle==2.8.0
pycparser==2.21
PySide6==6.3.0
PySide6-Addons==6.3.0
PySide6-Essentials==6.3.0
python-decouple==3.6
requests==2.28.0
shiboken6==6.3.0
toml==0.10.2
urllib3==1.26.9
